+++ py_slides_term/methods/rankers/lfidf.py
@@ -6,7 +6,7 @@ from ..rankingdata import LFIDFRankingData
 from ..data import MethodTermRanking
 from py_slides_term.candidates import DomainCandidateTermList
 from py_slides_term.share.data import Term, ScoredTerm, LinguSeq
-from py_slides_term.share.utils import extended_log10
+from py_slides_term.share.extended_math import extended_log10
 
 
 class LFIDFRanker(BaseMultiDomainRanker[LFIDFRankingData]):
