+++ py_slides_term/methods/rankers/mcvalue.py
@@ -4,7 +4,7 @@ from ..rankingdata import MCValueRankingData
 from ..data import MethodTermRanking
 from py_slides_term.candidates import DomainCandidateTermList
 from py_slides_term.share.data import Term, ScoredTerm
-from py_slides_term.share.utils import extended_log10
+from py_slides_term.share.extended_math import extended_log10
 
 
 class MCValueRanker(BaseSingleDomainRanker[MCValueRankingData]):
